day 5

+++ src/subjects/subjects.service.ts
@@ -1,7 +1,12 @@
 import { CreateSubjectDto } from './dto/create-subject.dto';
 import { UpdateSubjectDto } from './dto/update-subject.dto';
 
-import { HttpException, Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  HttpException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Not, Repository } from 'typeorm';
 import { PageOptionsDto } from 'src/common/pagination/page-option-dto';
@@ -21,29 +26,43 @@ export class SubjectsService {
     @InjectRepository(Grade) private repoGrade: Repository<Grade>,
     @InjectRepository(School) private repoSchool: Repository<School>,
     private gradeService: GradeService,
-
-  ) { }
-  async create(createSubjectDto: CreateSubjectDto, user: User): Promise<Subject> {
-    createSubjectDto.schoolId = user.school.id;
-    const school = await this.repoSchool.findOne({ where: { id: createSubjectDto.schoolId } });
-    const { name,gradeId } = createSubjectDto;
+  ) {}
+  async create(
+    createSubjectDto: CreateSubjectDto,
+    user: User,
+  ): Promise<Subject> {
+    createSubjectDto.schoolId = user?.school?.id || null;
+    const school = await this.repoSchool.findOne({
+      where: { id: createSubjectDto.schoolId },
+    });
+    const { name, gradeId } = createSubjectDto;
     if (await this.repo.findOne({ where: { name } })) {
       throw new HttpException('Tên đã tồn tại', 409);
     }
-    const grade:Grade = await this.repoGrade.findOne({ where: { id: gradeId } });
+    const grade: Grade = await this.repoGrade.findOne({
+      where: { id: gradeId },
+    });
     if (!grade) {
       throw new HttpException('Lớp không tồn tại', 409);
     }
-    const newSubject = this.repo.create({ ...createSubjectDto, name: `${name} ${grade.name}`, grade, createdBy: user, school: school ?? null });
+
+    const newSubject = this.repo.create({
+      ...createSubjectDto,
+      name: `${name} ${grade.name}`,
+      grade,
+      createdBy: user,
+      school: user.isAdmin ? null : school,
+    });
     return await this.repo.save(newSubject);
   }
 
   async findAll(
     pageOptions: PageOptionsDto,
     query: Partial<Subject>,
-    user: User
+    user: User,
   ): Promise<PageDto<Subject>> {
-    const queryBuilder = this.repo.createQueryBuilder('subject')
+    const queryBuilder = this.repo
+      .createQueryBuilder('subject')
       .leftJoinAndSelect('subject.grade', 'grade')
       .leftJoinAndSelect('subject.school', 'school') // Lấy thông tin trường
       .leftJoinAndSelect('subject.users', 'users'); // Lấy danh sách giáo viên phụ trách môn học
@@ -51,46 +70,45 @@ export class SubjectsService {
     const { page, take, skip, order, search } = pageOptions;
     const pagination: string[] = ['page', 'take', 'skip', 'order', 'search'];
 
+    // 🎯 Phân quyền dữ liệu
+    if (user.role === Role.TEACHER) {
+      const subjectIds = user.subjects?.map((subject) => subject.id) || [];
+
+      if (subjectIds.length > 0) {
+        queryBuilder.andWhere(
+          'subject.id IN (:...subjectIds) OR school.id IS NULL',
+          {
+            subjectIds,
+          },
+        );
+      }
+    } else if (user.role === Role.PRINCIPAL) {
+      queryBuilder.andWhere('(school.id = :schoolId OR school.id IS NULL)', {
+        schoolId: user.school.id,
+      });
+    }
+
     // 🎯 Lọc theo điều kiện tìm kiếm (bỏ qua các tham số phân trang)
     if (!!query && Object.keys(query).length > 0) {
       Object.keys(query).forEach((key) => {
         if (key && !pagination.includes(key)) {
-          queryBuilder.andWhere(`subject.${key} = :${key}`, { [key]: query[key] });
+          queryBuilder.andWhere(`subject.${key} = :${key}`, {
+            [key]: query[key],
+          });
         }
       });
     }
 
-    // 🎯 Phân quyền dữ liệu
-    if (user.role === Role.TEACHER) {
-
-      queryBuilder.andWhere(
-        '(users.id = :userId OR subject.created_by = :userId OR subject.created_by IS NULL) AND (school.id = :schoolId OR school.id IS NULL)',
-        {
-          userId: user.id,
-          schoolId: user.school.id
-        }
-      );
-
-      const subjectIds = user.subjects?.map((subject) => subject.id) || [];
-
-
-      if (subjectIds.length > 0) {
-        console.log(subjectIds);
-        queryBuilder.andWhere('subject.id IN (:...subjectIds)', {
-          subjectIds,
-        });
-      }
-    } else if (user.role === Role.PRINCIPAL) {
-      queryBuilder.andWhere('(school.id = :schoolId OR school.id IS NULL)', {
-        schoolId: user.school.id
-      });
-    }
+    
 
     // 🎯 Tìm kiếm theo tên môn học (bỏ dấu)
     if (search) {
-      queryBuilder.andWhere(`LOWER(unaccent("subject".name)) ILIKE LOWER(unaccent(:search))`, {
-        search: `%${search}%`,
-      });
+      queryBuilder.andWhere(
+        `LOWER(unaccent("subject".name)) ILIKE LOWER(unaccent(:search))`,
+        {
+          search: `%${search}%`,
+        },
+      );
     }
 
     // 🎯 Phân trang và sắp xếp
@@ -99,13 +117,13 @@ export class SubjectsService {
     const itemCount = await queryBuilder.getCount();
     const { entities } = await queryBuilder.getRawAndEntities();
 
-    return new PageDto(entities, new PageMetaDto({ pageOptionsDto: pageOptions, itemCount }));
+    return new PageDto(
+      entities,
+      new PageMetaDto({ pageOptionsDto: pageOptions, itemCount }),
+    );
   }
 
-
-
   async findOne(id: number): Promise<ItemDto<Subject>> {
-
     const example = await this.repo.findOne({ where: { id } });
     if (!example) {
       throw new HttpException('Not found', 404);
@@ -113,7 +131,12 @@ export class SubjectsService {
     return new ItemDto(example);
   }
 
-  async findOrCreateByNames(names: string[], gradeIds: number[], schoolId: number, user): Promise<number[]> {
+  async findOrCreateByNames(
+    names: string[],
+    gradeIds: number[],
+    schoolId: number,
+    user,
+  ): Promise<number[]> {
     if (!gradeIds.length || !names.length) {
       return [];
     }
@@ -127,14 +150,13 @@ export class SubjectsService {
     const school = await this.repoSchool.findOne({ where: { id: schoolId } });
 
     // Chuẩn hóa danh sách tên môn học theo `name + grade.name`
-    const formattedNames = grades.flatMap(grade =>
-      names.map(name => `${name} lớp ${grade.name}`)
+    const formattedNames = grades.flatMap((grade) =>
+      names.map((name) => `${name} lớp ${grade.name}`),
     );
-
 
     // Lấy danh sách môn học đã tồn tại theo `name + grade.name`
     const existingSubjects = await this.repo.find({
-      where: formattedNames.map(fullName => ({
+      where: formattedNames.map((fullName) => ({
         name: fullName,
         school: { id: schoolId },
       })),
@@ -144,28 +166,34 @@ export class SubjectsService {
 
     // Tạo Map kiểm tra nhanh (key: `name + grade.name`)
     const existingMap = new Map(
-      existingSubjects.map(subject => [subject.name, subject.id])
+      existingSubjects.map((subject) => [subject.name, subject.id]),
     );
 
-
     // Tạo danh sách môn học mới nếu chưa có
-    const newSubjects = grades.flatMap(grade =>
+    const newSubjects = grades.flatMap((grade) =>
       names
-        .map(name => {
+        .map((name) => {
           const fullName = `${name} lớp ${grade.name}`;
           if (!existingMap.has(fullName)) {
-            return this.repo.create({ name: fullName, grade, school, createdBy:user });
+            return this.repo.create({
+              name: fullName,
+              grade,
+              school,
+              createdBy: user,
+            });
           }
           return null;
         })
-        .filter(subject => subject !== null)
+        .filter((subject) => subject !== null),
     );
-
 
     // Lưu môn học mới nếu có
     if (newSubjects.length > 0) {
       const savedSubjects = await this.repo.save(newSubjects);
-      return [...existingMap.values(), ...savedSubjects.map(subject => subject.id)];
+      return [
+        ...existingMap.values(),
+        ...savedSubjects.map((subject) => subject.id),
+      ];
     }
 
     return [...existingMap.values()];
@@ -173,7 +201,9 @@ export class SubjectsService {
 
   async update(id: number, updateSubjectDto: UpdateSubjectDto) {
     const { name } = updateSubjectDto;
-    const exampleExits: Subject = await this.repo.findOne({ where: { name, id: Not(id) } });
+    const exampleExits: Subject = await this.repo.findOne({
+      where: { name, id: Not(id) },
+    });
     if (exampleExits) {
       throw new HttpException('Tên đã tồn tại', 409);
     }
@@ -183,21 +213,31 @@ export class SubjectsService {
     if (!example) {
       throw new NotFoundException(`Subject with ID ${id} not found`);
     }
-    const grade: Grade = await this.repoGrade.findOne({ where: { id: updateSubjectDto.gradeId } });
+    const grade: Grade = await this.repoGrade.findOne({
+      where: { id: updateSubjectDto.gradeId },
+    });
     if (!grade) {
       throw new HttpException('Lớp không tồn tại', 409);
     }
-    Object.assign(example, {name:updateSubjectDto.name, grade})
+    Object.assign(example, { name: updateSubjectDto.name, grade });
 
-    await this.repo.update(id, example)
+    await this.repo.update(id, example);
 
-    return new ItemDto(example);;
+    return new ItemDto(example);
   }
 
   async remove(id: number) {
-    const example = this.repo.findOne({ where: { id } });
+    const example: Subject = await this.repo.findOne({
+      where: { id },
+      relations: ['createdBy', 'school'],
+    });
+
     if (!example) {
       throw new NotFoundException('Không tìm thấy tài nguyên');
+    }
+
+    if (example.school == null || example?.createdBy?.isAdmin) {
+      throw new ForbiddenException('Không có quyền xóa');
     }
     await this.repo.delete(id);
     return new ItemDto(await this.repo.delete(id));
